Add files via upload
I have worked as a Web Developer for an extended period. One of my projects was to develop a Guessing Game with two other group mates by using HTML, CSS and JavaScript. The program asks for User's name, provides pictures and messages as hints, justifies answers, count number of times users answered for one question, prompts responses to users if they succeed or fail to answer the questions. It will give some makeup items if the user fails to respond more than three times correctly. At the end of the game, the program calculates the scores the players get.